Add rudamentary merging functionality.
This is basically able to do simple merging and value preservation.

I think the merging needs to altered so that it expects slices instead
of listed cells.

And the next point would be to add merging across multiple rows.

Add the support for embedded css and pandas

Now pandas can register this as a plugin for use
as an extension.

It also now allows embedding of arbitrary css in the output.

Add support for css embedding from file

This and a slight tweak to the name of the export function from to_file
to to_html to be consistent with pandas own.